+++ src/main/java/racingcar/Main.java
@@ -2,8 +2,8 @@ package racingcar;
 
 import racingcar.domain.Cars;
 import racingcar.domain.TrialCount;
-import racingcar.strategy.Generator;
-import racingcar.strategy.RandomValueGenerator;
+import racingcar.generator.Generator;
+import racingcar.generator.RandomValueGenerator;
 
 public class Main {
     public static void main(String[] args) {
