+++ src/main/java/racingcar/generator/Generator.java
@@ -1,4 +1,4 @@
-package racingcar.strategy;
+package racingcar.generator;
 
 public interface Generator {
     int nextValue();
