+++ src/test/java/racingcar/generator/ManualValueGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.strategy;
+package racingcar.generator;
 
 public class ManualValueGenerator implements Generator {
